Add ProjectManager
-Add project manager activity which shows a list of projects, and lets the user delete them along with all of its tasks
-Fix a bug where the list of tasks would not update after the task has been deleted

+++ app/src/main/java/hu/bme/aut/android/ktodo/data/todo/TodoItemDao.kt
@@ -14,7 +14,7 @@ interface TodoItemDao {
      * Gets all tasks from the given project.
      */
     @Query("SELECT * FROM tasks WHERE project_id = :projectId")
-    fun getTasksInProject(projectId: Int): List<TodoItem>
+    fun getTasksInProject(projectId: Long): List<TodoItem>
 
     @Insert
     fun add(todoItem: TodoItem): Long
